refactor: split GitHub quota response parsing
